Separate hmac test vectors and code
