+++ src/macs/hmac.rs
@@ -26,268 +26,3 @@ pub fn HMAC_SHA512(data: &[u8], key: &[u8], mac: &mut [u8; 64]) {
     m.update(data);
     mac.copy_from_slice(&m.finalize().into_bytes())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    struct TestVector {
-        data: &'static str,
-        key: &'static str,
-        mac: &'static str,
-    }
-
-    #[test]
-    #[cfg(feature = "sha")]
-    fn test_HMAC_SHA256() {
-        let tvs = [
-            // https://tools.ietf.org/html/rfc4231#section-4.2
-            TestVector {
-                data: "4869205468657265",
-                key: "0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b",
-                mac: "b0344c61d8db38535ca8afceaf0bf12b881dc200c9833da726e9376c2e32cff7",
-            },
-            TestVector {
-                data: "7768617420646f2079612077616e7420666f72206e6f7468696e673f",
-                key: "4a656665",
-                mac: "5bdcc146bf60754e6a042426089575c75a003f089d2739839dec58b964ec3843",
-            },
-            TestVector {
-                data: "dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd",
-                key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                mac: "773ea91e36800e46854db8ebd09181a72959098b3ef8c122d9635514ced565fe",
-            },
-            TestVector {
-                data: "cdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcd",
-                key: "0102030405060708090a0b0c0d0e0f10111213141516171819",
-                mac: "82558a389a443c0ea4cc819899f2083a85f0faa3e578f8077a2e3ff46729665b",
-            },
-            TestVector {
-                data: "54657374205573696e67204c6172676572205468616e20426c6f636b2d53697a65204b6579202d2048617368204b6579204669727374",
-                key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                mac: "60e431591ee0b67f0d8a26aacbf5b77f8e0bc6213728c5140546040f0ee37f54",
-            },
-            TestVector {
-                data: "5468697320697320612074657374207573696e672061206c6172676572207468616e20626c6f636b2d73697a65206b657920616e642061206c6172676572207468616e20626c6f636b2d73697a6520646174612e20546865206b6579206e6565647320746f20626520686173686564206265666f7265206265696e6720757365642062792074686520484d414320616c676f726974686d2e",
-                key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                mac: "9b09ffa71b942fcb27635fbcd5b0e944bfdc63644f0713938a7f51535c3a35e2",
-            },
-
-            // generated using: utils/test_vectors/py/main.py
-            TestVector {
-                data: "",
-                key: "",
-                mac: "b613679a0814d9ec772f95d778c35fc5ff1697c493715653c6c712144292c5ad",
-            },
-            TestVector {
-                data: "",
-                key: "f1586a3e5a4005f2c66a5a3252caf0c094db50f6ca6583233d5edfc68af1afe2fa5b19f512ea7c2b0f2ab32f63f144cc1fe29f505fe017bc1048c5457d2b0ba75451f1400062e3d21b261dbea786405899ceb8b4fe5554306c93a11115aec7de87168b3caabaed3b1e522624e5e603edf95fcafde92edc5c",
-                mac: "e3570a9092f615d3cbff259c257f77d7598adde77fbd9adc80c870af7652d49b",
-            },
-            TestVector {
-                data: "6ba58bd6fc7767fac07f92c082f244e4c5e65e18e223f41009fbaa18959f3f86275aa4adf5bbaba9d527117aad8d249e3f9da5b018e727e42dd7b9283179fbb16d0825151b2ea4b8c79b51aafbf5e4074c523c24c236c326a22e1d03827c15e06cef87259af1300e56bd847600225d292fdf8fec76b7892d8da76eac40227e81e48f50de9ec7c910944175978624923dd1c8f496bf8f603d5373213b77f52e2abc9d5a17e51da9b761ab2c2769388f119a7a54ef89b7943ac391986476cc78553aaacd914a55a99f96d7958c79e458b2e8d165657411e2e7d18af3d1f71b8ff95883a2ebe113c431e802e55ad7c9a4f811fe97a4d0d49723e625ef482d6e25200b220d3fbcbf922bd766fd1b1521fdafc01e57457b1bcb38b1edb96975ffeb24610198da5a68578c1928bde10909b70c577f8f8f19a1f7fedeb02473d8bf15fef7dd2619bf53366755dd8624113c21e68f8427420bdc79a4d965d608cdf9243bef73117c36d4283d99f5608d3b54bd50df34eb7642e0d049359c93b52a0b073e32b04bbe4a80ad03194d07e302b3c483f519f119583c12c7e5ab730cd6b4e8896d6a6b97705328d7b1abaaa73f2f72a05030c46ffbc535d47322c9f2e3adaf27e05579429be4681d0016d6981610c9b938c77056cb5f3e314e4cb4f26e8a4d2768eb6b6470451e1e7012573f58393c942d9037701b194ea7ef580ffd6192ba4ad833ee8a86efc93cc7e0db6d024e0acaadad322509f1d2039b9753f559ea7634bafd42c301c87e7f059a9322a58b62376248f2e650d5fef9b1688140f0197af2525fe1a840e9d3c496b2a1501907ca61b23da4636a66c2f5119f88ab5df4951a32a9d84bfb82aa4c28b676692ee6f08930132f54475e5b18bb4447fc063c630295892c577d2dd420f9a45f6fcda8205e5cde4c17420a59d5a3547ea69aa90b4f07f225f04b176c961f41cc39d386abad108ea34657c04982edfbd95f892bdf509f5392caba4beffd01dd7738d0cff29c83d70b36d49345c0e6fce7ab92a2dddbeda60ec15e9c035ffa51d22eae08e2e60d6d0b614b868bd9f798fc408aaabb4484669164cba34bb8ea91cc43883c49033b28c943fbe1f06731fb3c69aaa7d8a94a83d32aec91eccf311cc683562fc1528d1298ea0b6979e74096fd0fb18d6d70507091522f5219a62f57c3cc43819d0181eac2823b",
-                key: "",
-                mac: "dd0723502999d18d72958841ff7ac9e3e6f83ad037234641085309716499e3fe",
-            },
-            TestVector {
-                data: "fb4e6b25192afd7c2094c1bafe28a3af4d18a72db53eb072503c72a5296f4e037877707dcf216e7131ff7e4582ec2b6f8e643f8249d8a4a2d435d2aaab1f7af39247a954a687323b97a967fa3c7120db8c476966368d5607a51c24ca22caf903da1507d5b1ba824cd7780990c64f8870ac0db23ee5b5",
-                key: "731556f5ce9dc461b1863e",
-                mac: "2a75eaf2ea7898c25b666718bb99907c34efdb917e2c4b8a58de977630cc239f",
-            },
-            TestVector {
-                data: "6b2c95ac4d45a2f621c59e738e3604f7facc024620048908653b41fc2e0c5468354f6bdfda5c009aa2c9097ad72ed7fce0ac6d57728a2e99c11b9027fa079d88412d93a2bcbc4a18cddf55bfc318d812a4f07babf2fb407fd5ab5418a564386dab5ff446bad9024adf60768b613f90a7452d83832ab63b74d109efc3e276dc87b198f867f45d5d83db730591736f607c753dbda6fed7ffe01c7ad68da6da47050bd604582fc67b5141b179dcea6cb1f53a974f976af2748eb41189ec09743e160796244e29a0679bd202d079704b99abd428fe810e10e1a4f1d1cde0d0e80f8821e75b171d22ebd6650c7a5faf86615bbed4215f23ae6b06f30565c8ea7e9972ae0ee502d61e15cc32e43d999ef64c12dad3f4b2eb7b7c3641707ffe1fe435c34305f18f0fce2e4d1bb9623c7d2038e8e0618b78129974f91d54e4eca52d6e0b165ce487fc864ae1117fad35252062fe5a2099bf273e5f52cf3947c0b9152253c5c9b1ac21ef46d3ce5c40e4b4ce38cadf0623fb22a2ebef43ee3bab136c01d2013054a10a5130278bdd900c78b40774f5347588896c96827ea9e3a5430ba1f7d24cb2c299fc5521cef1a992701b01dd5403e898b515420c7f98a7cf7460776052d7177a6a92f9e0737d9556d6eab24c1d90d221aad6afc420d6f12427bbee2c5f69b30af714e182ae8c621aa0232194c4d8245f5d3474e30d72b0385197fb572562eb7984977153752950d129721fd72ca9bef832a2860ffb4f8d7809b1648a96f0ec84d476934a20da0d93a8620db1dbe2836a4c7f9fb7d21bb9aa0a5a3a9a8dee178e47acf2",
-                key: "6d800a770a093ad1c895cf6f3d7792af1777c138ac5eb0588a4f9dfe3de252731214a2806935428689a1b47a3782d51820c256c6ac8cd23f949e5f5b3b4848d5424a9826e23972161fb88a5c09738aa297e917c7cccdef53b9fe1aa02baa905fa1abcf076914fa900cbcce7f47ad9b78c424cb1a0bb751dc7faa59da5b9a5ea2a55a170778ee7c2cdcdba6497103acd3ff0f5d1e2846b4436c532963538db12e6a65aaca0eb6c5fddfa3b9f9817c124cb171a480901a261867f2e9039f1d3b16b280ec25597da1c34baca291012f6773c5",
-                mac: "c2cf5cb7bd2b822fccb2af564ec10ec66baf2ab34078a9805a82d2293b9c01a9",
-            },
-            TestVector {
-                data: "097c5939a88fa44fa13e467afa8927a39530cf64dbd00e721e71bfd8f4879f2931e27856a8897f56a50c264b79b077ebf926d4c4d7d1b7a93c512dafa76c97093e96ec30e6100d331c79b80f51333d9381c7fa5411ed79e0c7d3db9c1076f513e9063347087c04d2cd50782ca910f75cacea5db13d4d85aef178702dba31cf2512405d1bd187f3c535bf7f464c2feee3b618242d0e8ebd8f7e3ef7cf33e526119284a460b72b233ec9aad6607acf40850d784f85757d779497d3398270d478673534038495fb3c72604c1d44ccf95f74ef1d22ad469a08b5a14c5f698bfe35225431c3f0aa2738f6f62a02d7f1864f0f616ed1def6fabd500072add1e3b94b8284b47f971a0c3cfd2da81dec7e272aaa444e",
-                key: "e6a7652719123131ffe6458835d5ba40fcba1b0212ca33d2290a25e730b00f3ac9d28d59a883112eb0616feb113dd89770a13b5b9985619825d315b5df4d5c12e894b703647479e4cc96e5370e6dd8f4e110513665ce81ac951242278464583fe3ad8a7c6a1038d4108a4ae93eeee5b7a103f2ca658a5c3caa251daeca799c4e48e22be2826b293d9a76a084447e1e99e312d55cac0ac572c3ea7c8ac8ca0ff88d13c01ff7dffb2409cb56c04843a0b3c9308c838a91f0700694ab81fbb78b2fdac323e74ee1962d0e0e8732ae2c3783f18e65fd81cce77308c87758d6b28919a06ce260b126bcb9d097d30c00e76336f869860b421865c9c0fbb94f40dfcf62aa893dc74071fe640fae8c2dcaee7e9d53d5f5755b3f18ba1fc82839d435c92c60e551afa0cdfc3f8719e4031cf9912689bd1b55b727bf3c12d72def0d7e4a60caca70f64f75",
-                mac: "8f582f7f83513301b8ae03735284c122d87d58244f23ea3f43b48ecdaeabcb51",
-            },
-        ];
-
-        for tv in tvs.iter() {
-            let data = hex::decode(tv.data).unwrap();
-            let key = hex::decode(tv.key).unwrap();
-
-            let mut expected_mac = [0; 32];
-            hex::decode_to_slice(tv.mac, &mut expected_mac as &mut [u8]).unwrap();
-
-            let mut mac = [0; 32];
-            HMAC_SHA256(&data, &key, &mut mac);
-
-            assert_eq!(&mac, &expected_mac);
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "sha")]
-    fn test_HMAC_SHA384() {
-        let tvs = [
-            // https://tools.ietf.org/html/rfc4231#section-4.2
-            TestVector {
-                data: "4869205468657265",
-                key: "0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b",
-                mac: "afd03944d84895626b0825f4ab46907f15f9dadbe4101ec682aa034c7cebc59cfaea9ea9076ede7f4af152e8b2fa9cb6",
-            },
-            TestVector {
-                data: "7768617420646f2079612077616e7420666f72206e6f7468696e673f",
-                key: "4a656665",
-                mac: "af45d2e376484031617f78d2b58a6b1b9c7ef464f5a01b47e42ec3736322445e8e2240ca5e69e2c78b3239ecfab21649",
-            },
-            TestVector {
-                data: "dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd",
-                key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                mac: "88062608d3e6ad8a0aa2ace014c8a86f0aa635d947ac9febe83ef4e55966144b2a5ab39dc13814b94e3ab6e101a34f27",
-            },
-            TestVector {
-                data: "cdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcd",
-                key: "0102030405060708090a0b0c0d0e0f10111213141516171819",
-                mac: "3e8a69b7783c25851933ab6290af6ca77a9981480850009cc5577c6e1f573b4e6801dd23c4a7d679ccf8a386c674cffb",
-            },
-            TestVector {
-                data: "54657374205573696e67204c6172676572205468616e20426c6f636b2d53697a65204b6579202d2048617368204b6579204669727374",
-                key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                mac: "4ece084485813e9088d2c63a041bc5b44f9ef1012a2b588f3cd11f05033ac4c60c2ef6ab4030fe8296248df163f44952",
-            },
-            TestVector {
-                data: "5468697320697320612074657374207573696e672061206c6172676572207468616e20626c6f636b2d73697a65206b657920616e642061206c6172676572207468616e20626c6f636b2d73697a6520646174612e20546865206b6579206e6565647320746f20626520686173686564206265666f7265206265696e6720757365642062792074686520484d414320616c676f726974686d2e",
-                key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                mac: "6617178e941f020d351e2f254e8fd32c602420feb0b8fb9adccebb82461e99c5a678cc31e799176d3860e6110c46523e",
-            },
-
-            // generated using: utils/test_vectors/py/main.py
-            TestVector {
-                data: "d74021fce1a1a5175e9463c6453f69fb8017f100e28e76cde969b17d58b98ad6cb3df9d60254b048a594c81a7242e7ad8f9030b9827acc786d3e12b0a28372f4d4ecfd7791751311dbc4e4499318114cb643d23cb5fcf80e8940a355c98512995166c20a92bbb4c50dc366e42af8ad0634a1a95029ccf392b91796bf1772e65909ff117e01c19229651ee8bb7979bbf9b9a4d87e740687bfd224e0ec5587d39be8f9d4279f8849a18a863a5cb71f1daf6db9771ddff6811bcac2a6935ca35b269c06bfa2405490bb831733336c1139940a8375f490108c7bda9777c3f41febea14c227e1301447a2ef430171d6bbb0e4ef7ba8aeaa74c2d6b402d575195f27340d0e3bcd1d3c6b787c3b93de19ecf17e9d093408dee20def2eb63a758f7a4fc0f6cd13c22a884772a16593a8595d23783c38eaf3dd83fb154770697165d16f841620f93572156f64dedd35c56af34d209c091736dd134cbbdf444c92152e4c3bed864812343ed97c3fc30d6bb9aaa6a4d7113abde534843ad9d1d4cf590f7ae57358698a026f9e9a1c00608d23bebcccc3c82ce916fbcc061500dcd9f8a436cf16f76b7ef2f48bbe8b5dec11fbeee22b63c1e285370a48fc7fcdaf73dabee32df7bca4876cb99ccc715aa98f61d73d9aeede3c5b1b901fc094be7c859dd01364ea68f03bcb04af2e939d4bf5468f5fe1e97570014ad0d8913a35a155c9ec3bc68481e676411da547c0d530e14ce60aee88e72ab56a9e6598502fbbef85691eaea7f1034b104e675bc84b19cc53b5c2b701a00d7542bd1027452ced8396610edc561e0ec1c7e2d382fb5e4f6fa3bb0eabcb0cc7beb27ec642d6ce54a8a0cd85",
-                key: "775d9904",
-                mac: "c6eda655b90dbdea90b1bf3f9040154fcd3cbdc286bdf10163d7c46922c33fbba676be43fa9c2cd08bc1ce507d591fb5",
-            },
-            TestVector {
-                data: "42469412c7ab75c8de1164ddfade1cb667ae76d7bd5ff30d189add7d1473677366bba0583793b365f1f6a40ccea5db703ab2d1a7cc3c0641504bd34abfd21c8712e4222d3da102abed15a4ea5ea1874baa5969403a8b190fe557427cc5bf4b8002e32c5c7906b9d55bccf8dc560c34c26ac527994bff2246fae017f687e7918296de53f3e9a658afd610671fc469e19ca8c63df55f42d35c1e55622afb667b8781647ae973334530a7cf3d2995ba79a70375a29bf8de4e14fb89bddb8fc707c4612033088e6e3311504d4880c0e45c508e9cd69147cbbb24665fb77f815ea4118d6dc56107cea3e2fca4c469547fe5ed485f5bd9d9c53c3f901077cc4727fa4f754a2468e238cb54bf3ec29adb2fbd1a5ac04478fb9c658d90ec0d6d3970e7a965e1e5fe6c68b89b52c289d872215d6e5ef6ddf15f1f8a75f01aea4bc93660b3c7992bb7cac4f69f237938c4b46000f267a99912fc7b86312a889d535984a98221ee598bc7cef8105f520dd9fafb69a884ac54b6208b7770482b36866f9fe1c543a83b5b51bf74fdacec8bc027e36e7457a581b8c76e63bacea505ed96bb30fe439b60b2122db8d01ce39776de814ffbc2f6e36e0b2f9e50b3e364cf1c24e18a5d3a1a6143c6eb68d26176dad45d4b25cb2d4de839f2b3eeda2822aff5ed8e715707917fa5fc4fa754d16a0dc4612075333be752ea9c5535044bb87f0dee3fca428227ff2e191cf9cdbe7b82253cf89dec5cb977e15bf63515bb7e92a565f843e3c2eceefdc5271644aba7be9e15af054e9999e87278c277c9fd382ec0f75c56b5b8800aec6e9efc7b8c43d128483b5d5b5207ab3ffcd2f313202dbcf3c59387373fc6826aa3fdc8deb55e04d0659cd4e33ed81225f440d1edc0d5fa0ca9089f548a1cc1632491c48d87b489d4353ba2a050d9b51d46ca5f947822621b1435568237881b2dedeb5f",
-                key: "e929a5d6ebaf779611c0614cd6bee29b123ce13560e581751d921c7d43c9038fde877bff54d13ff3f624169f532b1a720ca1272a37103891fcb40638dfb2e64fc58fad1a461a851cf41da8d493def737ebe85f284a6b4d7c5a5314c453c42e830ab4c292d9a4c4ae3b43b019a10b17f78a53d78ff1621b0302221726c6d70ddb6d83ccebe74aef775b63f73116cd4740b8328ccae8c20627017bac904ceeb64cf6accb189e63c1ee674b4e697b09879e825a5239c84d9641893e329e8ea4c93e9c35b0a4e0ebab63cc6c4eb4e114dbb6c4030c8d0b82c6eeaf0644bdae61512b384b0f0fb2a9ebcd1813abc7c3659a730d2db1a06dab2934592ebca98cc9acc6d7aabfa2049e1c42721c6617a7ef5123db078215939a7d537ff0e65186ea59a2b4911373b9661e558a15301c16b542aacb6391a9ec4b156d7f079cc6542938ee624e89ea8dced036a314e44368bd93e254a6df721dc4797ce040045b6a0642a8357639ce031a1e98a2a6e12855e7654dc48f2979387498ed5778de8369dd2c8716d991d13c4cb00dfc919c28cc09ecc2257a7ed02bc14e9092a1c7bdc7437cbedc1d4bbdae107880f8170652f7e42657355e418f237bca062f7dfcbcfcec035ea8cc5228de92b972b1b90acbbce9b3212d3c5465a99ffb2866c891ebc75264ab52517e453ab19364d99c7a7d0946d39cb79fada1a788a47ccfaf86f8cbc94950a11a20713ef911bd47c16ce1dc4b29393fad79b4c2d6d1559303ca2ccc116ae7e1d58ac3efc5828f1b976f05efc4faa0db1e44891f76d2f0b951067a384e5557525d39f673f59e5ffea93a8b032f63ef0accc8abf1018c27a97d40f8c71fad4bd34e385bf1caa53c8ea7afe00af4903b5f5c65b4ced7d62bc0f78c01f58248b15d7cf4a33048df78975795e3275e8ccbed0e1d853bf3812cdd169dabcc78947851585702eb8924d4cabb7b3ced27d3357c08733ffa4e22fdc72b5b211edf9e0add3c9ca732fe7a9cbb95dbd646fb09c5e99db56a67a52d4915ca00766e324089dcd2c9d988b5d40cfe0c0b27e132daa30165870c6b7408e829ea188b21d312bbd18319e292122117a09a0b8bc377352f77406ec7fee553d5c325c8ce2ae3a2e6603f6e201e4520fb8d3109fd4637e70b6223b45c1783eb76fb17cb46f5a6efe2c82108e63b871478ac289cc8a7da76ef06fb2191eeed76154a2013c1693cd2ad4ffd3c6f6e09b5",
-                mac: "8e7c66c7fd9aa83f979559f46ab97b84a4451e01c96a5a76266bf1c8471894489c325431212d3710c79da61a3b53645c",
-            },
-            TestVector {
-                data: "d2641e10df0d1941bfbbd0abc1f61ce1d5ac980cb5be4e685687",
-                key: "03fed297f3bf88e3db03c9ed7efb93be3dcd79fedd1c694116096dd98e9d592e90fcfa7f38eb424717f20d1d1208c1ca79744daa9c0cd8909ce034af6066cb5ab9b6daaa4127353061459488311e3bee478c8c79e711d84613842b7d2140eb85f7309ce0a3511fbb42f1cb68537b4d618fdd6ff5ce0dd7f8552883f0bb549868136000b2549e943493f68b813b5a6357fa7b62bcd4627b7f21bb0a5ca970b4a90c28d02f6440a14f0453a9ac7f8eccdec55656e086c02bfc488ebbdafe107c7cfc43e982687b4a9a9b8b7148328559f0360a4a22d05e5d3fdece30914b424ddc0edc4543ffa0db34275e3cb944fd7a805abe5ed29c9eb217c4b95d2ef8bd37ba086060778868e4b177ebe0ee56e840c9c6c48de5881fed4bda455599e4e0c92d847a9de43752f3e200c13be93e7b4ec28dcbef1e4535c01b95b971dee45aa0338f64e765831324037dd21b9314fab26dd506ff1a43b278b85553e316f182c9504b63d3cbba3868cf20d930edfa79b7d0e8ddf2257166c2195837c942b0d24144733110fc11da194e9f2d7047056596dbe92d0e0afab9d262c304ede32fa5b2b7004b6044ad02473649def29c4aec8abc3e2fd986822ecedf1e70d96cdc2df18c6363d47860973e95175435d9bb2393fc8f5e1b2813ae52452aa2e9c08a48479425aea4c46419ec2fbaefe8a061288aca7a8952f76a799ad7cfb2182c42d9bf611b70e67fc8f88a460f75ba3b5b630e357caadf70eeb2d56c7ce212dcaf8d750b20facf05e3ec941658bd947e4b7dfded583ba1e50fbcdb5cc31e9298c9ef42fafd50740e51e0603ca2e4d8c9f1d0fcaeff934fd3eaa7e2b049aae2d36fe059a2999658cbbab4a582816ef284801a87531178a4922497bda990112e3070a5db77ef973edb60ac1b7dd88292bc595c822d1283da7ec0e1d5f02dede84bd87b99d9f035a8cb5a24a097f7e7d6258a237b382dbc6c553b5b9f54b7f82e4e85f68465149f8d999a99e8e431de8987fe07b032716d1a0dcba33e1fb2b7c72da0f63e55556ee6db0c1a9e90693f9115e7e49542711e441c11b2ed35465afc8802b8f8a4311c609ffe45a63c9f7341f3a6f7e3925294624e7c2112",
-                mac: "0e06f2b681c8f8775cc3dccb4ef575eda3d6e39576b0dcb5bca33fef696d643cf6c5593f8e31ea09e1208cac0feaaa28",
-            },
-            TestVector {
-                data: "ebad2b3f39f51545a87f9fde91559f0669fb1894047f9e78b14a694f479d4b67edc94d593fd270b39aaa72e231e2964507d6a6b3421304d238ff5b0a93b9826dabed66d6817d5ef04bb95c91b54e3619724fff025280ce53fee70030672d5eb0be5d58879b1725716d32ab5ba6ceb142ea88502c171fa4679570e96c929d29968c2637f3749757d4fff31048552a277e7e700698654597770a9f34d864bef78165f76f44fb6898cf09c4d6405ee3fe0da1969a7fd0a2529881ee72dab4ae2e677ec1a8319e70be9faeeefc6f5e5b1f60cf45c8197f6b4b452272f9062027e28f46381256abf4e0768021a9f5905b842d85540d2346f3f66ccd27faa9c5622a467845b7f21920a7cfeddb78a846138346cb4cc697",
-                key: "04021597cf80d1796be68a95b760b28cfe789dc3c697e37eb3b721497fa50d9ff8cc88b3b78dbdec6e75a2235c685c719534d1650c733fe19413fc234ecd23121daa8eafd1e0e88c783128f2aa4a5d93f44b85a33ec96fac86190fa0393fb4867f4c213d4a6e5b3c021c30e8671fdd1e973c2dfbd35f455d296e901965cbdff567bc71a55238bfaa2783c7d14e262a45074f128b053c8a6f564d186a241de404b193f7b9301f44e6561dbba17c0cb0ec443f973a17e4ffc2d96a08ff6df9bfcc87da143a230075aa9151b8fba618a847774a251f2aba84809de664b81dec30492b6b51c3b4d99aa1244b4d8ed23be8aff1694d974fbb268e8b145cc4ec2651abcf0e218af944fe717892590a1dbe04aec329678a12864d7ace6e7ede996aff98366e2e2a0b1136b522595cc42714b6c40261490f5bc9eb46fe72dd28b77446865d45597154b18ac214263ece614b7cffac944abcf68894a6182181",
-                mac: "58bf6cc62528891fde44fb20e39f6231d988e9f02fcfcb29bd0c867b500d25ad08590ee0cd2167dc028522fecb8c19bf",
-            },
-            TestVector {
-                data: "6029f207326b02b14b71056efb52731eb67ff8b5ef565d2d62ed828cd6a36db4910d55e7cecb1f99130e0e99bde63b7124d1e2f18220aa60f2bf922cbc2c4f279e162a2e7ba40f6e528929dc79b12b4bdbbbd0d5d0f92cb18cf742a4e55c071f1830161919a057beaa84828f5bf8ab5db1a80ab2f8e06540b9314dbf2b109c4c0d9fe231d9b3b02b14c065d6bb396f1805150cb80e3dc9847cc59fe4dc09a732defa5159ee912ea68e5705da81edace26922057eda5127e23e90e706863ae852ef883bc6a4fb2568c498771fe8dc700c939177955c2511a5ebc23328e7ae269f8eda156073d4c873a88a97d932727a1d5560d90540d08dbaa04a85fe16ccfeed1af730786497af524c436b55148aecaae05c38cec6da8875f36460ba156f346d3f5f6d47eb9a31bfb69fb93a61bef5f9312a8a1ae759f45135d3799c8e5aba06d2b656bca18f7d9497cf8a8ac5a1f60cc5643017c2631e1fbef448caa55e2344df715feed0ea206d00b0be1fffe2aa008b6a60da7242d747f0092ac9a5880645b7e284682be86bd08cd3bcfff6698bfa6d2bececb55a26cc87efc65e69984bf841e4f6e57a8d01e21c1caddd49fb7939aec7d45a4267e9688dd5a9a564ce30d48d46ecdbd0147272da7891b17fad122fa80180e9c00adfebb22b0a6f4f83fe499f7222d32c2da04caacdce0520e31e5b488730c48dba68d70aa52e2d80b90ce19db230653941c7a546eee09db32ed90375dfe6edd0099caad97f539f0e8678103953747e67a6560e33a15d35b09b8ce7b5a9b5747e1a0321b2b6989cde67851480ef369e0420fbcf4611a717eee9ea76f1ecc5796e4f9497c6e7a1877d04418626fd9c83b1a548b608e7b2f15c194d0ac246500ec81b279c5ca34d9229dc9a36a5f47884ef5df4e27f7f95d7045591a9d2c037cf3e3b750fcf31f3cc1a3e2e4132894cb5085947774408098a1369fb2d70dc90a1d8466204457eba6f1e0b107b9a418b563b488d1ca342ec2692639383c54a0b9f90cb21bbc9c6b4339aac18d74f0e67336493b77524d6",
-                key: "985f0e28fafc3c21c02d39c18aa39f990b381b16b11ca22cd385c070095eb6e9c422d3c8f15d9df6755a242d14cec6a279886ef3e6f15b74497d11cb53a637ae64d0ffba0eac092e24fe1a847419ab63cb00d0fc0a0c20d136561c2f6663c0258d337daa6d9270829b7596ff3f8457d69d277f063768ef697fb3de85bf626c7d27a01b3726c1bed75ab501c4945d487d7bc488d644a7b81bc69f045eeda1e9bb30653e8e072ca7bb02c97af70ae340fa8c9cc3ec21d9e0263f41ccea99d1e36991a6f5ed152d76ab8c2c11e13ac1fa784d52ada49e2bbbc20149bf9afce9e9f9cfb07022db615992ea14dd405a4f7445bc6b9a6c13a4d9537ba13a680581c31041b9af8a63a9c216440ca0b774d46794e90f8b55cdac16780d36ed42b344dea9ea26058841b21575e23c80aaf187e33f12386ee3b299c1e2286055bff2144078de41aab6bf807ebddeed811e8e11c545a5e32bc216f64dc5f60304eea7d1147fab83e48568ac82010b68b35e90b2c9808b83e4f97c8f827d597e87562a5fb6009d2cac0ebea314187019baae6b0aa7b5d113cd8501763e05f2c5b8507574bfeea210e435b2b4cc75839cd32ae4d0755e0c2a509c6b75b785941c8507de5a3286cf61d6e681ed4c70a19c0b56c19cfdca7f381c514b21209d05b492f3925bcd981fbfaacbd82b3d526822",
-                mac: "03dddb231e54253ee9b404a25ef68213ab6dc88f369e3440f03d5b62ffc6c8ad7ce480f7a170f61a4ddd8f9ac1b276ad",
-            },
-            TestVector {
-                data: "deca016c789111e4833d6c69f79c8121fe64918ee5fded215324c763b95a2e183387a5c6a14708556a8dd5725d9dbb77d0ddab77ed3244e33672c87f52cdf5d742bf3b7a07f5a268f9588efb09b0e82ef7475f62eeb48dd92a20c13863048184d94e4353094ced29de4b2ecc397ca0d962533786adae2956862c1e4067d3248b87c3aa248dac9501321b690c41cda16a9b6c4d1b318000387d81259d20b12b2fc0515ea559f02cbd1cbe6a45a5b30be12effa8f30433d0edf1d3c9fc82aed260d1dc6e53a778f516f2de1666e13bbe1321d211b80b63295856ee300f852cda83561044cdd472c9b5add7fcc0f9f6acb8254d0dc7cc5a80254ab1178662a96fee8bd14a3b591b2c08505a43a9b95a92d06b5730ee60dbd39fa70ed017e6590f952eabd9f36de743824571d1579b34979dc8520cdf1d4a843953ef89f479578e798324108de47e8ec7d9c85cf90a863b9f08794fb35bb342472356c9e36ad4040523fe75fa0afb563448a3f533e3e3a27d5fb20db4fae3d4cbcfbff30b6b0ca872c8de91e0ea3fb14e24f9827882378df3f8f78ce7b76ad74ffa8ff7cbf6cadac2578e34ac97f297892f28c6aa9f926a1a7e5e768a01a752a3532ac878f92ea941931de3e7ad462cefa965de90c0d87e8ee3d509ddb321a57db99149bb7859d3c4611ed28dff95533d85555f0df234b5b1",
-                key: "c5158d52be36",
-                mac: "8f2414f7a851a4d96b009fa56ff9930d77a8961cc86c6cbdda9f5d47e03bea2b32df77f326b905eac1ce3346c494eabe",
-            },
-        ];
-
-        for tv in tvs.iter() {
-            let data = hex::decode(tv.data).unwrap();
-            let key = hex::decode(tv.key).unwrap();
-
-            let mut expected_mac = [0; 48];
-            hex::decode_to_slice(tv.mac, &mut expected_mac as &mut [u8]).unwrap();
-
-            let mut mac = [0; 48];
-            HMAC_SHA384(&data, &key, &mut mac);
-
-            assert_eq!(&mac, &expected_mac);
-        }
-    }
-
-    #[test]
-    #[cfg(feature = "sha")]
-    fn test_HMAC_SHA512() {
-        let tvs = [
-            // https://tools.ietf.org/html/rfc4231#section-4.2
-            TestVector {
-                data: "4869205468657265",
-                key: "0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b",
-                mac: "87aa7cdea5ef619d4ff0b4241a1d6cb02379f4e2ce4ec2787ad0b30545e17cdedaa833b7d6b8a702038b274eaea3f4e4be9d914eeb61f1702e696c203a126854",
-            },
-            TestVector {
-                data: "7768617420646f2079612077616e7420666f72206e6f7468696e673f",
-                key: "4a656665",
-                mac: "164b7a7bfcf819e2e395fbe73b56e0a387bd64222e831fd610270cd7ea2505549758bf75c05a994a6d034f65f8f0e6fdcaeab1a34d4a6b4b636e070a38bce737",
-            },
-            TestVector {
-                data: "dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd",
-                key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                mac: "fa73b0089d56a284efb0f0756c890be9b1b5dbdd8ee81a3655f83e33b2279d39bf3e848279a722c806b485a47e67c807b946a337bee8942674278859e13292fb",
-            },
-            TestVector {
-                data: "cdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcd",
-                key: "0102030405060708090a0b0c0d0e0f10111213141516171819",
-                mac: "b0ba465637458c6990e5a8c5f61d4af7e576d97ff94b872de76f8050361ee3dba91ca5c11aa25eb4d679275cc5788063a5f19741120c4f2de2adebeb10a298dd",
-            },
-            TestVector {
-                data: "54657374205573696e67204c6172676572205468616e20426c6f636b2d53697a65204b6579202d2048617368204b6579204669727374",
-                key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                mac: "80b24263c7c1a3ebb71493c1dd7be8b49b46d1f41b4aeec1121b013783f8f3526b56d037e05f2598bd0fd2215d6a1e5295e64f73f63f0aec8b915a985d786598",
-            },
-            TestVector {
-                data: "5468697320697320612074657374207573696e672061206c6172676572207468616e20626c6f636b2d73697a65206b657920616e642061206c6172676572207468616e20626c6f636b2d73697a6520646174612e20546865206b6579206e6565647320746f20626520686173686564206265666f7265206265696e6720757365642062792074686520484d414320616c676f726974686d2e",
-                key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                mac: "e37b6a775dc87dbaa4dfa9f96e5e3ffddebd71f8867289865df5a32d20cdc944b6022cac3c4982b10d5eeb55c3e4de15134676fb6de0446065c97440fa8c6a58",
-            },
-
-            // generated using: utils/test_vectors/py/main.py
-            TestVector {
-                data: "",
-                key: "",
-                mac: "b936cee86c9f87aa5d3c6f2e84cb5a4239a5fe50480a6ec66b70ab5b1f4ac6730c6c515421b327ec1d69402e53dfb49ad7381eb067b338fd7b0cb22247225d47",
-            },
-            TestVector {
-                data: "",
-                key: "87c163ef979f07e7b2dc08f6fb993a01ae78adb78801ff33737062fe8ddd07538152e15416802c7a3da2068a1dcf4536beca2a24b99c2a1051c6b69c91f1b25aafeb32442c66e7bd272e7db4f439b831d81817d82536aec5830923fb71db42a635241d347150e3696fb5b857a0480f4e0982e32a409bf278fd765dd137fd92c6fe6aff791fc808c0c0908d9b19e60774d381f74a6f75ae6a6ab19e85ef43371cd501f34b98eb4e03f22dac6f9411b51ecd83437a2bcf88ea2c07a609651987c2fe2f9b848155c2abc2c7fcf91ff63c6ed0c1c0cd5cacd0224342f7fa44a760e40b0447dfe35d4f0a5ec9667364a6df40bc32a28f9feb599a1c8787501fc572d2f757257878519b693c611b7914961b46e2da016e529d5d936caf3600bad317ba6d77d87220b313b4ba6f090aa702852cc15fb2b08832bf8bd102f1c14ce7450fd753e7330077c8cf313497b77762d1a457278386b8962ff4e6265ceb71d321bd891f138a48505bac43d9ea8eb9ba894ef9038ba9ba138b3f7e1da17a0b6ecd064b9b0ae23270797ce40a4e2089bf467a3d0e0026c866e940f2d47443409284",
-                mac: "8cf8a53352178b036ff4f117be0771a15a9e6bdab7d299c29ad8290c894bd90690eb092f1fdccaece260443ef2e40d30ccdea78fe2af52bb96ce872403b7078f",
-            },
-            TestVector {
-                data: "79e5a8782508cd17f02fa545636ac5d53bdf8c6b5aa8f59964cc59bff7a5c97bb1e811f2c7abf905db25b02be0b096fdeec5119f202c94ee68e0dab9913049eb7f65abb83bb23f8b2c4e87b25b5d55c64d013ad0f15eefa41bef9cd02bed1ff1037f1c7dbd5357ab96d4592e26e784bb87d4911b62fad9f34f2c9d9b37b4d90f3397090ce4ac2775c145e779a2765f1b4a16155ddd69e2f164e71e7f4ebaf2ccfc1dd3d46780a8d885480034d071d3bd23741e7c8b7036c0ecc332a966d0f3fd6b2d4446aaa79f941bdcb6ee38bd82aed2b7534a8777f77702a68451fb0c8fcf53d2336f6244e38cc12a8aecf2895833cfa259363bb8e15cff3cd4d3bb65f454c604e641dd5ffc2dde61dc96b572a4716c4158cbfa7a4bd8e37c9c6f54d391647c360b5b9a66ce348fcc288964b18b081e8bfddb66eb5cc1fbd860cca5b35cef3391bb479e736d79282e5ad2cb86236a3d32e4ed28ed2f0a7691",
-                key: "",
-                mac: "6adce509e70a5b96f733a695935b5c4c7f06387c2fdf54919ca7367be9c30b04a9e47821f669da432b2e4174a4042bb97e1632f76d52fb47f0b79e937dd479e4",
-            },
-            TestVector {
-                data: "96670c0b307009c533402174258de0a0c404119774c23e7394bf04466c70d38aa8fd8ea44909f96bc7612ec834282eb7071024e0b6e5c1ad8a7b1c971a049e12382dc371d25370be8d597a31ba565c2aaa453ec26d776b062860d806fd989f1b3557a7d439321f843e7883fc8d4633186005348c30775228485fc649fc96c206e5f260c08a2852d0cd2c573f752edb1de973c8f0406a3c46a81ed80f70fa1ceabf9bdd3cd84cd71084c8d3e49d73e32993e8ac195cb425bf636d5651335da71be591537db5a5868101f3b885d1f37a9217ce410380765fab03d7b5cce6106b93ffde0a74fe72b9f73e9be52931c5a4da3ae3ba4f26280e69ac9b7d0454f583e09aecb36c578bef77be54c8e88384cbf37d4bf82bbf1218510e52b6b671e64c4ad7c9a5b9f39430d99313790fee4cc160f04e7112246c9ebc488ec25f072f49da444108b74692b965a2acf147a57fee7ee8aa762640d4cf9419f8c32c456e5f4ec580513697495d347e4c955792cb9fe26b9d2c8731e771387da3018893881274e7761b3a23ab5f12",
-                key: "f9f6253fd4956d43b1be5d29a42b0052874a2ae98b0c310e6acc907a650cd6c538031e1bfffa3e4ebc3009fda7be21f40728df14e1f3c3ea94d0413f7f4a7318a6e887c73cc25c027e579df69dd8f56ea2ba7d3944e28640e52140f831810cf9c293e0feb0af9dd13413da9f2cf303aff1b0fc4bac8095fe9d14a3611f53bfaa7cda6f92f6f78800506a0d36611955800896683a315a7801fed5304575eaa52d8b2cac71fbc12a2d206d2c73cf94b0468920b746b587c52a3fdcbb474f4a12b17c32841c866fd8d6a8c9c1edd84f0ebbd1b7fff0cea5b7b9bf29f69c21ea5991dd4bd6ee86293158cae650ee14dd59745f99eac1218ae257f06d46c4ff7ddbfe469e9fdd222ca21c3de7ec7e290d3ba09a727aaf69af96ecfe9cc4964f205e2fb6c138d8f51736b0a4454f4ab160923e9fe521ee92717fc07405d2d19659b53718442b7abcb35f2c1f8c41e94578359f17a0e09ff6b830aa347594c2edb1573df96158ae89f93b5d1e04d18674cf32f638099fa26a311ddd1503fe5a30485515bf87e669e67616396fff50bc1430ae2d77323d0b0fe7856f548f79224c5e62b07fedf5fb44951134ef9cde5bcbc6b603317f470fede13209bbdced596c9668ff2dd281528fdfddb178622d025482fa7acc70b78a1976601296e148cc0b7d18dcd775de51809785a774a455ff691233783eb3150974793e0ebdb833dfbf730ac5cfd01199953218198f498338929887faabeffc6bd56e80ce46923a5e37cebfac41218bea9d6dbe3d",
-                mac: "ece7334f7ab5e740385a68c57ddf7c09a7d273d06b51e66a7a630bd88ff0e85c086201dfaeee969fd987c8f263d14b6a677c8d97067ab6251849f0faffb89b42",
-            },
-            TestVector {
-                data: "3542b6e7fb2929405453a5d6de1128eeb74f5a5a5b89c1862ecc94b27078d09533bc2d529a2cce234bd7bdb8ffb06069ea7e1e3c24cc571012795fbe6629f6a4478eb988cc2194546a9fc1c61f1f659f9ee0088f78dd7222b0cf132cfe7eee74526e7e505c9868b02a0435e7ae93fd3b23c3168b24d56d7608d3a8ffd2e394fd616ad01b7bf66be0d45a240bdac13c1d5dd96ac7a127218d88004d58755b72dbeb842d51a7071963b5a020f8d813545244d835ddf27e8968bad5ce72d87817ea46a425e2a2756d5077bb2f58156d90285be8e857dc7f88779c67c9f57ccafd6de9ab7d0b4dbb109ea9e1cefc",
-                key: "53ca77a40384dda4d14564cd5aafebee904ffe3895cb2bb70d29bfd62cbee0c2bbcd4f5463890d5bb8d8186877b088a4e58796cedb0895742a95a841afb9ce2e676c8b6a5b1f22580c49f04de5f2753157f59c2c96dac1146a299cd7616789b59440a3111a87fa1ea71c81ff5d8c71787dcae630f54ff249d1ceb47c31925a84e9690ae4f6f74e960e492d05fbeff48f86dc5f86e09e7531dd9cd2096b81b950778ef9b58a2564c65fa217023ec7cbe2879ef5969d05efd9e242c19c2051c6c3832b1dafabf655f8eacfe897dde422a0f4f09031ba6385cece8fae40c5c5e64bf8a7eb882bc64f85b4eb9574e163574d0b69f88ad1138fc8bbe4702d909a9fbe660e1719127e5cdc527f67ebd86510eefc38093dcec135f88cde94d32a9e13f8781188c6e4277c25cfd4e34a919db5487401035c1f58f288576cb8f90b761935199351f97be12f95c6b0be3e3865b3c989c72a99d0a83be6e3c01aa5d936ba512521f8401728e3bef9ee051e90cf7cc4d4feed2d275498b89a5e8a3420607c401465e18dce5bc65658eb34e2fbdfc728a61be6a329f4681121ac153927e9dbe6dbe4d649c2bde4623de88eb3cb1398a266d4dc03b8d9517f30e4775d34d61763f2b74867301fa794d82c0d7a327088653dc9465fee129a84938f45b1d0e224792e8f9a707df02edaaa40dfd5bb315b19bf0320ea822545a36c3c1302cb6cc5782873c31994fb81fef2",
-                mac: "0bf04a5005be5e33c9203033647c54c814440b0eb3d3a587d83792879df87d59b85229bfbc39f46130a6b70e7b35efab0f9b595f522522625d6fe09eb07af023",
-            },
-            TestVector {
-                data: "72873821e4fe7486c461041e89cd09aa9c7d4228d6230496343d29d53e859284347ea56f896ca0d469ec6e42b1ea2f1dc5eada8a130bc31deea092b6d79f9d4b15b99203fc176476196f562706350f0ee622c10d52894558d78ba98381224f319ce6088b41a8177ca1f486f36ed3c6ea918337c23cebd5166ecb3f4dd0deac976cd4db1c6f60bf616e442cf5e50fd0514465ce095e9c356238fada17a40311126b69168c355bbc7c7aafd088c0e2cff9c61854e31ed2a9b829e9e005badb716ea2bf012b0b0abb82bc0d86cd87d0b50fd8f6ee367747742ae376a076a2995644c103d78b6812490cccaf020569c42c4427b995c997c20edd741f7c591632dea31d9fc8adee6d36562b2b1d8a37943dca24e9d20642ce789b2cfb3c117c79ec5d0e658f4d4a952dcc301c9700cd85da78757e3a463f19d206f13320dc8e2d11a45ffe389566abe59ca48fef5d00e228197d23908c78fe41cb2cc52b22c12f2345628a9c6db13e84067dac449ced30c05ee38789f1b6bb095f0788cb9c80d9afb70b7d70731e6c95ac454f9d2e96573feb0c44c3625bcb8a54c150299e238d27fe522499df3bb0e0f93e32925656c56d5f61da2d6a266d14425b42bd938ad4bd1511543656e01aa7add9d592bd69ea9a88522d1b7c57caa1161c8bb9a9328dc65ed6f092f07f7d0ce052516a951588b85f2c3186f4d7881f1e6ed773b154514d89d0fd1cba09fba9b6a72d916679a0b2df6edfea99f7506f8edae949428584ee7661d0f5025060b4820cb5c053aef9fb893861b0782ee0441527ddc478137b9f5dbca892a28dfc919064cc67159e8bc5a5038891a2a74a45cecdba60f2d5a672bacf275445c44e6390f6c05c7beafeaeb156c136a307694177818c9defc49d37772e801db1b1ff12e04edbe16e991de4ebce8e4d788bd2f4328390015d924dc7572f546429e246e19de6422bbce256ade9ed6ce4a908d7a3acab0a82dacf104a7762717bb2254613654d3047702fbc99528998c184b33b262a6eaf44f8f7281d269f4cb713325a26c8ac4cec1463bdd722bd6c48e9c32205fee9cfda6d0ecb023ea2cd4b22b39a39d843e15919c47e85d760617fef0e23a735915d30fbe01c359ecf28725b26580ed9810e8119c0f214854b7f99fce37aed0e4e6c193eeb9c9613d1066c1f744ba4277fc4affae80fedc8a84b9c73e43da519b301e58529f31a440838846870dd2acc4961b1d5d04eb6f12b8dbc15bae1ed5103ecb49a5e3b69db3812ead3fd4b6ce62e50546f6bbd3b17c6b6c9c1b22da928268baba7c2d0fde541b59e62",
-                key: "d0d66fbb250e7fb9c3282da15f3558c8050ff28041",
-                mac: "91c000e62404b58f2d9ff35570709e786ac9bc14a55d768e6d9caeb7a1c0baaf9b3d453fd9da27392901502b86b83e7b6511aedd6684c5c8f34bcdd3f19fa306",
-            },
-            TestVector {
-                data: "606fcba357c006b3feefbdb8108114c271400c7dd00ce2c48c8d7a3cfe20fd285f8e5abe4036547e2b9e8cffed00354b6bbc0702bf5c9a1de4c2295e3ebced9a5210ed7f85babbfc6b556b01be31b8b14ffbc949802cfa05283beb90f40b7512c4b406d4b7ab30413a475bcf2362bbdb35a8ab7ef3f8a7e4ed7bd47615d7ff616f65f7a2ff26fa3530af68f85b6a760c347377a429d2e9fa021082e611e3b47912a3c1348d5065817f34b5d0696fe57b652830c671bc600891965d75601e1a81002503e152f745b3d058fb6f0a635050584e1c1c80a65e4e4893d47f36dbe35e9f97438277080f89adf5c0ddc3f111bc72627a235e49f3a1c78f52a0777cf2e2153b0ca695e5653122a915b454fe9da577e73f3a3c9778b4ecc3b61f61e27a7eaeb19b794c71ebc4dc13a8ff4795f6333286d0f70dc8c5c54a71421f66d047f7030d3977c6bc803eeccc9c94e60fe595164ac18ff932326b4392a7fc371608b723ef15df31248b871cbefa9a3656b79a97e75ac827220627098b6d39156bfd0e8aeb2f70d8b9f9f2e1d39815ef7396a5b217c32b26c5231b28f3d8ca13d638cdaf1b0d9593a21ae8806af6800aecc27a9d3127322011646c76cfea05b85983ebfc4efa1ebfe25eaddc5534c673e260dd6bfb55640219598f7e4e8bdc9b68585d89a94cc63f91a3e0d183a4c6bcc63b00ca55d16dd48fd43feb35f2fde9a947c2ebfc5d16453766d8531d1fbff8366f316b872a3101c012ce76674e07e51ed488b5cdc9cce3863931a63307c8ce7eb4fac414b3a2c74b6c2390b3bef9232cad5037019af9d700edcf3de0fc5dd64c66b65b8ad62db6d8c1af3301545798b5e34db2caf9befbd2383025eb7fcba5bf314b90ae2f02792a18df4c085d6934d0462e017002ab91520be6e0b176ca8249c9925b68dcc08df9731841863d9ecc492f6075e4e0aca3234a1ec30a4224068837b7fc624858b8fa81c63801e2b20b0f1ebe770717630000646ca9d2d297e3c56004fa5d9fe6bd9352bc7ad72e4554192c5bfe20d1d80ceb45bcb22c59d1dfa95351791ef217e6b6bd21c19a91a263f6f070d1b4eba75fc227408232b895b14b247c165b210b740a9ec6203a9703be91fa2b0cfab1cccacbdea8b5d6fd70eba8d100fc0393a3bd8fa4868d2539caf28f222528b238b26e160490a248466a10a073f24f952edfd3ae244812052f0346d63d149d2ad11fcb70c8ddf530477547e3379741e972934a63b2857a901df25104d69c149ff8f26f9978655dc2247492f0411e2e67c100e8739ec0a8b037062f0ffd63cf9d30a6504cca9cbf7ec7dffc93014308a1d638924ad09cd07668",
-                key: "9d29e43f7134c989f5b40ffb61bb4778c8fa95d9f8b27e4b0c2db33bc55a602c180550d2366b3c4f6a181bb83b046c214201f098a41baeb06ef5e95327aa7373486cd54af721ad47593bc95e9c2055f02f41492fc762e7dcf5e5e9010036a188c1b7023b59f1f6563a7ff47e1399a70ae05ab4b9203a7aa2e1c0030fb143b6d96ab3907a07c50ec9d69ba7b61a2c4038b49ffed3711b827dc683cf1b27ff39742b91911d7d45beba5ecea944d96c6397aa8f489bf875c64b57a8dbc543b23c3d23d5d34b6dd37b95ba7a537f83f45fb395b0275049b61a85fd5956f593d8fc3c32497c556479d52a277b5f93d573e07ef63675576ab9c8e84389dd82929bb259613b615b72e924a0c7715479a809b1541f8d0380b4dfe535d4b8c2bdc9440fa3858c3434c01a0b4b16a3a83c589d9db9b1bcde2438cd5f0a244e7879640ec51f18ce1d7219fb50ff843f1073d7f33fc421314cc3284e7d21a8fa161fa8181b1283645e78ab27dcee0965458ac8de4cac5558359c5d78c330146bfa",
-                mac: "11008133980ee7fad23844028a6a367baa9733e4f3e9b47c03b94b35fa99648c0bc4706b232728d8b76149c985a6a547d36f9349afb0b2043a5227b64e45337a",
-            },
-        ];
-
-        for tv in tvs.iter() {
-            let data = hex::decode(tv.data).unwrap();
-            let key = hex::decode(tv.key).unwrap();
-
-            let mut expected_mac = [0; 64];
-            hex::decode_to_slice(tv.mac, &mut expected_mac as &mut [u8]).unwrap();
-
-            let mut mac = [0; 64];
-            HMAC_SHA512(&data, &key, &mut mac);
-
-            assert_eq!(&mac, &expected_mac);
-        }
-    }
-}
