+++ tests/fixtures/hmac_sha256.rs
@@ -1,0 +1,65 @@
+[
+    // https://tools.ietf.org/html/rfc4231#section-4.2
+    TestVector {
+        data: "4869205468657265",
+        key: "0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b",
+        mac: "b0344c61d8db38535ca8afceaf0bf12b881dc200c9833da726e9376c2e32cff7",
+    },
+    TestVector {
+        data: "7768617420646f2079612077616e7420666f72206e6f7468696e673f",
+        key: "4a656665",
+        mac: "5bdcc146bf60754e6a042426089575c75a003f089d2739839dec58b964ec3843",
+    },
+    TestVector {
+        data: "dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd",
+        key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+        mac: "773ea91e36800e46854db8ebd09181a72959098b3ef8c122d9635514ced565fe",
+    },
+    TestVector {
+        data: "cdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcd",
+        key: "0102030405060708090a0b0c0d0e0f10111213141516171819",
+        mac: "82558a389a443c0ea4cc819899f2083a85f0faa3e578f8077a2e3ff46729665b",
+    },
+    TestVector {
+        data: "54657374205573696e67204c6172676572205468616e20426c6f636b2d53697a65204b6579202d2048617368204b6579204669727374",
+        key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+        mac: "60e431591ee0b67f0d8a26aacbf5b77f8e0bc6213728c5140546040f0ee37f54",
+    },
+    TestVector {
+        data: "5468697320697320612074657374207573696e672061206c6172676572207468616e20626c6f636b2d73697a65206b657920616e642061206c6172676572207468616e20626c6f636b2d73697a6520646174612e20546865206b6579206e6565647320746f20626520686173686564206265666f7265206265696e6720757365642062792074686520484d414320616c676f726974686d2e",
+        key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+        mac: "9b09ffa71b942fcb27635fbcd5b0e944bfdc63644f0713938a7f51535c3a35e2",
+    },
+
+    // generated using: utils/test_vectors/py/main.py
+    TestVector {
+        data: "",
+        key: "",
+        mac: "b613679a0814d9ec772f95d778c35fc5ff1697c493715653c6c712144292c5ad",
+    },
+    TestVector {
+        data: "",
+        key: "f1586a3e5a4005f2c66a5a3252caf0c094db50f6ca6583233d5edfc68af1afe2fa5b19f512ea7c2b0f2ab32f63f144cc1fe29f505fe017bc1048c5457d2b0ba75451f1400062e3d21b261dbea786405899ceb8b4fe5554306c93a11115aec7de87168b3caabaed3b1e522624e5e603edf95fcafde92edc5c",
+        mac: "e3570a9092f615d3cbff259c257f77d7598adde77fbd9adc80c870af7652d49b",
+    },
+    TestVector {
+        data: "6ba58bd6fc7767fac07f92c082f244e4c5e65e18e223f41009fbaa18959f3f86275aa4adf5bbaba9d527117aad8d249e3f9da5b018e727e42dd7b9283179fbb16d0825151b2ea4b8c79b51aafbf5e4074c523c24c236c326a22e1d03827c15e06cef87259af1300e56bd847600225d292fdf8fec76b7892d8da76eac40227e81e48f50de9ec7c910944175978624923dd1c8f496bf8f603d5373213b77f52e2abc9d5a17e51da9b761ab2c2769388f119a7a54ef89b7943ac391986476cc78553aaacd914a55a99f96d7958c79e458b2e8d165657411e2e7d18af3d1f71b8ff95883a2ebe113c431e802e55ad7c9a4f811fe97a4d0d49723e625ef482d6e25200b220d3fbcbf922bd766fd1b1521fdafc01e57457b1bcb38b1edb96975ffeb24610198da5a68578c1928bde10909b70c577f8f8f19a1f7fedeb02473d8bf15fef7dd2619bf53366755dd8624113c21e68f8427420bdc79a4d965d608cdf9243bef73117c36d4283d99f5608d3b54bd50df34eb7642e0d049359c93b52a0b073e32b04bbe4a80ad03194d07e302b3c483f519f119583c12c7e5ab730cd6b4e8896d6a6b97705328d7b1abaaa73f2f72a05030c46ffbc535d47322c9f2e3adaf27e05579429be4681d0016d6981610c9b938c77056cb5f3e314e4cb4f26e8a4d2768eb6b6470451e1e7012573f58393c942d9037701b194ea7ef580ffd6192ba4ad833ee8a86efc93cc7e0db6d024e0acaadad322509f1d2039b9753f559ea7634bafd42c301c87e7f059a9322a58b62376248f2e650d5fef9b1688140f0197af2525fe1a840e9d3c496b2a1501907ca61b23da4636a66c2f5119f88ab5df4951a32a9d84bfb82aa4c28b676692ee6f08930132f54475e5b18bb4447fc063c630295892c577d2dd420f9a45f6fcda8205e5cde4c17420a59d5a3547ea69aa90b4f07f225f04b176c961f41cc39d386abad108ea34657c04982edfbd95f892bdf509f5392caba4beffd01dd7738d0cff29c83d70b36d49345c0e6fce7ab92a2dddbeda60ec15e9c035ffa51d22eae08e2e60d6d0b614b868bd9f798fc408aaabb4484669164cba34bb8ea91cc43883c49033b28c943fbe1f06731fb3c69aaa7d8a94a83d32aec91eccf311cc683562fc1528d1298ea0b6979e74096fd0fb18d6d70507091522f5219a62f57c3cc43819d0181eac2823b",
+        key: "",
+        mac: "dd0723502999d18d72958841ff7ac9e3e6f83ad037234641085309716499e3fe",
+    },
+    TestVector {
+        data: "fb4e6b25192afd7c2094c1bafe28a3af4d18a72db53eb072503c72a5296f4e037877707dcf216e7131ff7e4582ec2b6f8e643f8249d8a4a2d435d2aaab1f7af39247a954a687323b97a967fa3c7120db8c476966368d5607a51c24ca22caf903da1507d5b1ba824cd7780990c64f8870ac0db23ee5b5",
+        key: "731556f5ce9dc461b1863e",
+        mac: "2a75eaf2ea7898c25b666718bb99907c34efdb917e2c4b8a58de977630cc239f",
+    },
+    TestVector {
+        data: "6b2c95ac4d45a2f621c59e738e3604f7facc024620048908653b41fc2e0c5468354f6bdfda5c009aa2c9097ad72ed7fce0ac6d57728a2e99c11b9027fa079d88412d93a2bcbc4a18cddf55bfc318d812a4f07babf2fb407fd5ab5418a564386dab5ff446bad9024adf60768b613f90a7452d83832ab63b74d109efc3e276dc87b198f867f45d5d83db730591736f607c753dbda6fed7ffe01c7ad68da6da47050bd604582fc67b5141b179dcea6cb1f53a974f976af2748eb41189ec09743e160796244e29a0679bd202d079704b99abd428fe810e10e1a4f1d1cde0d0e80f8821e75b171d22ebd6650c7a5faf86615bbed4215f23ae6b06f30565c8ea7e9972ae0ee502d61e15cc32e43d999ef64c12dad3f4b2eb7b7c3641707ffe1fe435c34305f18f0fce2e4d1bb9623c7d2038e8e0618b78129974f91d54e4eca52d6e0b165ce487fc864ae1117fad35252062fe5a2099bf273e5f52cf3947c0b9152253c5c9b1ac21ef46d3ce5c40e4b4ce38cadf0623fb22a2ebef43ee3bab136c01d2013054a10a5130278bdd900c78b40774f5347588896c96827ea9e3a5430ba1f7d24cb2c299fc5521cef1a992701b01dd5403e898b515420c7f98a7cf7460776052d7177a6a92f9e0737d9556d6eab24c1d90d221aad6afc420d6f12427bbee2c5f69b30af714e182ae8c621aa0232194c4d8245f5d3474e30d72b0385197fb572562eb7984977153752950d129721fd72ca9bef832a2860ffb4f8d7809b1648a96f0ec84d476934a20da0d93a8620db1dbe2836a4c7f9fb7d21bb9aa0a5a3a9a8dee178e47acf2",
+        key: "6d800a770a093ad1c895cf6f3d7792af1777c138ac5eb0588a4f9dfe3de252731214a2806935428689a1b47a3782d51820c256c6ac8cd23f949e5f5b3b4848d5424a9826e23972161fb88a5c09738aa297e917c7cccdef53b9fe1aa02baa905fa1abcf076914fa900cbcce7f47ad9b78c424cb1a0bb751dc7faa59da5b9a5ea2a55a170778ee7c2cdcdba6497103acd3ff0f5d1e2846b4436c532963538db12e6a65aaca0eb6c5fddfa3b9f9817c124cb171a480901a261867f2e9039f1d3b16b280ec25597da1c34baca291012f6773c5",
+        mac: "c2cf5cb7bd2b822fccb2af564ec10ec66baf2ab34078a9805a82d2293b9c01a9",
+    },
+    TestVector {
+        data: "097c5939a88fa44fa13e467afa8927a39530cf64dbd00e721e71bfd8f4879f2931e27856a8897f56a50c264b79b077ebf926d4c4d7d1b7a93c512dafa76c97093e96ec30e6100d331c79b80f51333d9381c7fa5411ed79e0c7d3db9c1076f513e9063347087c04d2cd50782ca910f75cacea5db13d4d85aef178702dba31cf2512405d1bd187f3c535bf7f464c2feee3b618242d0e8ebd8f7e3ef7cf33e526119284a460b72b233ec9aad6607acf40850d784f85757d779497d3398270d478673534038495fb3c72604c1d44ccf95f74ef1d22ad469a08b5a14c5f698bfe35225431c3f0aa2738f6f62a02d7f1864f0f616ed1def6fabd500072add1e3b94b8284b47f971a0c3cfd2da81dec7e272aaa444e",
+        key: "e6a7652719123131ffe6458835d5ba40fcba1b0212ca33d2290a25e730b00f3ac9d28d59a883112eb0616feb113dd89770a13b5b9985619825d315b5df4d5c12e894b703647479e4cc96e5370e6dd8f4e110513665ce81ac951242278464583fe3ad8a7c6a1038d4108a4ae93eeee5b7a103f2ca658a5c3caa251daeca799c4e48e22be2826b293d9a76a084447e1e99e312d55cac0ac572c3ea7c8ac8ca0ff88d13c01ff7dffb2409cb56c04843a0b3c9308c838a91f0700694ab81fbb78b2fdac323e74ee1962d0e0e8732ae2c3783f18e65fd81cce77308c87758d6b28919a06ce260b126bcb9d097d30c00e76336f869860b421865c9c0fbb94f40dfcf62aa893dc74071fe640fae8c2dcaee7e9d53d5f5755b3f18ba1fc82839d435c92c60e551afa0cdfc3f8719e4031cf9912689bd1b55b727bf3c12d72def0d7e4a60caca70f64f75",
+        mac: "8f582f7f83513301b8ae03735284c122d87d58244f23ea3f43b48ecdaeabcb51",
+    },
+]
