+++ tests/fixtures/hmac_sha512.rs
@@ -1,0 +1,70 @@
+[
+    // https://tools.ietf.org/html/rfc4231#section-4.2
+    TestVector {
+        data: "4869205468657265",
+        key: "0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b",
+        mac: "87aa7cdea5ef619d4ff0b4241a1d6cb02379f4e2ce4ec2787ad0b30545e17cdedaa833b7d6b8a702038b274eaea3f4e4be9d914eeb61f1702e696c203a126854",
+    },
+    TestVector {
+        data: "7768617420646f2079612077616e7420666f72206e6f7468696e673f",
+        key: "4a656665",
+        mac: "164b7a7bfcf819e2e395fbe73b56e0a387bd64222e831fd610270cd7ea2505549758bf75c05a994a6d034f65f8f0e6fdcaeab1a34d4a6b4b636e070a38bce737",
+    },
+    TestVector {
+        data: "dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd",
+        key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+        mac: "fa73b0089d56a284efb0f0756c890be9b1b5dbdd8ee81a3655f83e33b2279d39bf3e848279a722c806b485a47e67c807b946a337bee8942674278859e13292fb",
+    },
+    TestVector {
+        data: "cdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcd",
+        key: "0102030405060708090a0b0c0d0e0f10111213141516171819",
+        mac: "b0ba465637458c6990e5a8c5f61d4af7e576d97ff94b872de76f8050361ee3dba91ca5c11aa25eb4d679275cc5788063a5f19741120c4f2de2adebeb10a298dd",
+    },
+    TestVector {
+        data: "54657374205573696e67204c6172676572205468616e20426c6f636b2d53697a65204b6579202d2048617368204b6579204669727374",
+        key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+        mac: "80b24263c7c1a3ebb71493c1dd7be8b49b46d1f41b4aeec1121b013783f8f3526b56d037e05f2598bd0fd2215d6a1e5295e64f73f63f0aec8b915a985d786598",
+    },
+    TestVector {
+        data: "5468697320697320612074657374207573696e672061206c6172676572207468616e20626c6f636b2d73697a65206b657920616e642061206c6172676572207468616e20626c6f636b2d73697a6520646174612e20546865206b6579206e6565647320746f20626520686173686564206265666f7265206265696e6720757365642062792074686520484d414320616c676f726974686d2e",
+        key: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+        mac: "e37b6a775dc87dbaa4dfa9f96e5e3ffddebd71f8867289865df5a32d20cdc944b6022cac3c4982b10d5eeb55c3e4de15134676fb6de0446065c97440fa8c6a58",
+    },
+
+    // generated using: utils/test_vectors/py/main.py
+    TestVector {
+        data: "",
+        key: "",
+        mac: "b936cee86c9f87aa5d3c6f2e84cb5a4239a5fe50480a6ec66b70ab5b1f4ac6730c6c515421b327ec1d69402e53dfb49ad7381eb067b338fd7b0cb22247225d47",
+    },
+    TestVector {
+        data: "",
+        key: "87c163ef979f07e7b2dc08f6fb993a01ae78adb78801ff33737062fe8ddd07538152e15416802c7a3da2068a1dcf4536beca2a24b99c2a1051c6b69c91f1b25aafeb32442c66e7bd272e7db4f439b831d81817d82536aec5830923fb71db42a635241d347150e3696fb5b857a0480f4e0982e32a409bf278fd765dd137fd92c6fe6aff791fc808c0c0908d9b19e60774d381f74a6f75ae6a6ab19e85ef43371cd501f34b98eb4e03f22dac6f9411b51ecd83437a2bcf88ea2c07a609651987c2fe2f9b848155c2abc2c7fcf91ff63c6ed0c1c0cd5cacd0224342f7fa44a760e40b0447dfe35d4f0a5ec9667364a6df40bc32a28f9feb599a1c8787501fc572d2f757257878519b693c611b7914961b46e2da016e529d5d936caf3600bad317ba6d77d87220b313b4ba6f090aa702852cc15fb2b08832bf8bd102f1c14ce7450fd753e7330077c8cf313497b77762d1a457278386b8962ff4e6265ceb71d321bd891f138a48505bac43d9ea8eb9ba894ef9038ba9ba138b3f7e1da17a0b6ecd064b9b0ae23270797ce40a4e2089bf467a3d0e0026c866e940f2d47443409284",
+        mac: "8cf8a53352178b036ff4f117be0771a15a9e6bdab7d299c29ad8290c894bd90690eb092f1fdccaece260443ef2e40d30ccdea78fe2af52bb96ce872403b7078f",
+    },
+    TestVector {
+        data: "79e5a8782508cd17f02fa545636ac5d53bdf8c6b5aa8f59964cc59bff7a5c97bb1e811f2c7abf905db25b02be0b096fdeec5119f202c94ee68e0dab9913049eb7f65abb83bb23f8b2c4e87b25b5d55c64d013ad0f15eefa41bef9cd02bed1ff1037f1c7dbd5357ab96d4592e26e784bb87d4911b62fad9f34f2c9d9b37b4d90f3397090ce4ac2775c145e779a2765f1b4a16155ddd69e2f164e71e7f4ebaf2ccfc1dd3d46780a8d885480034d071d3bd23741e7c8b7036c0ecc332a966d0f3fd6b2d4446aaa79f941bdcb6ee38bd82aed2b7534a8777f77702a68451fb0c8fcf53d2336f6244e38cc12a8aecf2895833cfa259363bb8e15cff3cd4d3bb65f454c604e641dd5ffc2dde61dc96b572a4716c4158cbfa7a4bd8e37c9c6f54d391647c360b5b9a66ce348fcc288964b18b081e8bfddb66eb5cc1fbd860cca5b35cef3391bb479e736d79282e5ad2cb86236a3d32e4ed28ed2f0a7691",
+        key: "",
+        mac: "6adce509e70a5b96f733a695935b5c4c7f06387c2fdf54919ca7367be9c30b04a9e47821f669da432b2e4174a4042bb97e1632f76d52fb47f0b79e937dd479e4",
+    },
+    TestVector {
+        data: "96670c0b307009c533402174258de0a0c404119774c23e7394bf04466c70d38aa8fd8ea44909f96bc7612ec834282eb7071024e0b6e5c1ad8a7b1c971a049e12382dc371d25370be8d597a31ba565c2aaa453ec26d776b062860d806fd989f1b3557a7d439321f843e7883fc8d4633186005348c30775228485fc649fc96c206e5f260c08a2852d0cd2c573f752edb1de973c8f0406a3c46a81ed80f70fa1ceabf9bdd3cd84cd71084c8d3e49d73e32993e8ac195cb425bf636d5651335da71be591537db5a5868101f3b885d1f37a9217ce410380765fab03d7b5cce6106b93ffde0a74fe72b9f73e9be52931c5a4da3ae3ba4f26280e69ac9b7d0454f583e09aecb36c578bef77be54c8e88384cbf37d4bf82bbf1218510e52b6b671e64c4ad7c9a5b9f39430d99313790fee4cc160f04e7112246c9ebc488ec25f072f49da444108b74692b965a2acf147a57fee7ee8aa762640d4cf9419f8c32c456e5f4ec580513697495d347e4c955792cb9fe26b9d2c8731e771387da3018893881274e7761b3a23ab5f12",
+        key: "f9f6253fd4956d43b1be5d29a42b0052874a2ae98b0c310e6acc907a650cd6c538031e1bfffa3e4ebc3009fda7be21f40728df14e1f3c3ea94d0413f7f4a7318a6e887c73cc25c027e579df69dd8f56ea2ba7d3944e28640e52140f831810cf9c293e0feb0af9dd13413da9f2cf303aff1b0fc4bac8095fe9d14a3611f53bfaa7cda6f92f6f78800506a0d36611955800896683a315a7801fed5304575eaa52d8b2cac71fbc12a2d206d2c73cf94b0468920b746b587c52a3fdcbb474f4a12b17c32841c866fd8d6a8c9c1edd84f0ebbd1b7fff0cea5b7b9bf29f69c21ea5991dd4bd6ee86293158cae650ee14dd59745f99eac1218ae257f06d46c4ff7ddbfe469e9fdd222ca21c3de7ec7e290d3ba09a727aaf69af96ecfe9cc4964f205e2fb6c138d8f51736b0a4454f4ab160923e9fe521ee92717fc07405d2d19659b53718442b7abcb35f2c1f8c41e94578359f17a0e09ff6b830aa347594c2edb1573df96158ae89f93b5d1e04d18674cf32f638099fa26a311ddd1503fe5a30485515bf87e669e67616396fff50bc1430ae2d77323d0b0fe7856f548f79224c5e62b07fedf5fb44951134ef9cde5bcbc6b603317f470fede13209bbdced596c9668ff2dd281528fdfddb178622d025482fa7acc70b78a1976601296e148cc0b7d18dcd775de51809785a774a455ff691233783eb3150974793e0ebdb833dfbf730ac5cfd01199953218198f498338929887faabeffc6bd56e80ce46923a5e37cebfac41218bea9d6dbe3d",
+        mac: "ece7334f7ab5e740385a68c57ddf7c09a7d273d06b51e66a7a630bd88ff0e85c086201dfaeee969fd987c8f263d14b6a677c8d97067ab6251849f0faffb89b42",
+    },
+    TestVector {
+        data: "3542b6e7fb2929405453a5d6de1128eeb74f5a5a5b89c1862ecc94b27078d09533bc2d529a2cce234bd7bdb8ffb06069ea7e1e3c24cc571012795fbe6629f6a4478eb988cc2194546a9fc1c61f1f659f9ee0088f78dd7222b0cf132cfe7eee74526e7e505c9868b02a0435e7ae93fd3b23c3168b24d56d7608d3a8ffd2e394fd616ad01b7bf66be0d45a240bdac13c1d5dd96ac7a127218d88004d58755b72dbeb842d51a7071963b5a020f8d813545244d835ddf27e8968bad5ce72d87817ea46a425e2a2756d5077bb2f58156d90285be8e857dc7f88779c67c9f57ccafd6de9ab7d0b4dbb109ea9e1cefc",
+        key: "53ca77a40384dda4d14564cd5aafebee904ffe3895cb2bb70d29bfd62cbee0c2bbcd4f5463890d5bb8d8186877b088a4e58796cedb0895742a95a841afb9ce2e676c8b6a5b1f22580c49f04de5f2753157f59c2c96dac1146a299cd7616789b59440a3111a87fa1ea71c81ff5d8c71787dcae630f54ff249d1ceb47c31925a84e9690ae4f6f74e960e492d05fbeff48f86dc5f86e09e7531dd9cd2096b81b950778ef9b58a2564c65fa217023ec7cbe2879ef5969d05efd9e242c19c2051c6c3832b1dafabf655f8eacfe897dde422a0f4f09031ba6385cece8fae40c5c5e64bf8a7eb882bc64f85b4eb9574e163574d0b69f88ad1138fc8bbe4702d909a9fbe660e1719127e5cdc527f67ebd86510eefc38093dcec135f88cde94d32a9e13f8781188c6e4277c25cfd4e34a919db5487401035c1f58f288576cb8f90b761935199351f97be12f95c6b0be3e3865b3c989c72a99d0a83be6e3c01aa5d936ba512521f8401728e3bef9ee051e90cf7cc4d4feed2d275498b89a5e8a3420607c401465e18dce5bc65658eb34e2fbdfc728a61be6a329f4681121ac153927e9dbe6dbe4d649c2bde4623de88eb3cb1398a266d4dc03b8d9517f30e4775d34d61763f2b74867301fa794d82c0d7a327088653dc9465fee129a84938f45b1d0e224792e8f9a707df02edaaa40dfd5bb315b19bf0320ea822545a36c3c1302cb6cc5782873c31994fb81fef2",
+        mac: "0bf04a5005be5e33c9203033647c54c814440b0eb3d3a587d83792879df87d59b85229bfbc39f46130a6b70e7b35efab0f9b595f522522625d6fe09eb07af023",
+    },
+    TestVector {
+        data: "72873821e4fe7486c461041e89cd09aa9c7d4228d6230496343d29d53e859284347ea56f896ca0d469ec6e42b1ea2f1dc5eada8a130bc31deea092b6d79f9d4b15b99203fc176476196f562706350f0ee622c10d52894558d78ba98381224f319ce6088b41a8177ca1f486f36ed3c6ea918337c23cebd5166ecb3f4dd0deac976cd4db1c6f60bf616e442cf5e50fd0514465ce095e9c356238fada17a40311126b69168c355bbc7c7aafd088c0e2cff9c61854e31ed2a9b829e9e005badb716ea2bf012b0b0abb82bc0d86cd87d0b50fd8f6ee367747742ae376a076a2995644c103d78b6812490cccaf020569c42c4427b995c997c20edd741f7c591632dea31d9fc8adee6d36562b2b1d8a37943dca24e9d20642ce789b2cfb3c117c79ec5d0e658f4d4a952dcc301c9700cd85da78757e3a463f19d206f13320dc8e2d11a45ffe389566abe59ca48fef5d00e228197d23908c78fe41cb2cc52b22c12f2345628a9c6db13e84067dac449ced30c05ee38789f1b6bb095f0788cb9c80d9afb70b7d70731e6c95ac454f9d2e96573feb0c44c3625bcb8a54c150299e238d27fe522499df3bb0e0f93e32925656c56d5f61da2d6a266d14425b42bd938ad4bd1511543656e01aa7add9d592bd69ea9a88522d1b7c57caa1161c8bb9a9328dc65ed6f092f07f7d0ce052516a951588b85f2c3186f4d7881f1e6ed773b154514d89d0fd1cba09fba9b6a72d916679a0b2df6edfea99f7506f8edae949428584ee7661d0f5025060b4820cb5c053aef9fb893861b0782ee0441527ddc478137b9f5dbca892a28dfc919064cc67159e8bc5a5038891a2a74a45cecdba60f2d5a672bacf275445c44e6390f6c05c7beafeaeb156c136a307694177818c9defc49d37772e801db1b1ff12e04edbe16e991de4ebce8e4d788bd2f4328390015d924dc7572f546429e246e19de6422bbce256ade9ed6ce4a908d7a3acab0a82dacf104a7762717bb2254613654d3047702fbc99528998c184b33b262a6eaf44f8f7281d269f4cb713325a26c8ac4cec1463bdd722bd6c48e9c32205fee9cfda6d0ecb023ea2cd4b22b39a39d843e15919c47e85d760617fef0e23a735915d30fbe01c359ecf28725b26580ed9810e8119c0f214854b7f99fce37aed0e4e6c193eeb9c9613d1066c1f744ba4277fc4affae80fedc8a84b9c73e43da519b301e58529f31a440838846870dd2acc4961b1d5d04eb6f12b8dbc15bae1ed5103ecb49a5e3b69db3812ead3fd4b6ce62e50546f6bbd3b17c6b6c9c1b22da928268baba7c2d0fde541b59e62",
+        key: "d0d66fbb250e7fb9c3282da15f3558c8050ff28041",
+        mac: "91c000e62404b58f2d9ff35570709e786ac9bc14a55d768e6d9caeb7a1c0baaf9b3d453fd9da27392901502b86b83e7b6511aedd6684c5c8f34bcdd3f19fa306",
+    },
+    TestVector {
+        data: "606fcba357c006b3feefbdb8108114c271400c7dd00ce2c48c8d7a3cfe20fd285f8e5abe4036547e2b9e8cffed00354b6bbc0702bf5c9a1de4c2295e3ebced9a5210ed7f85babbfc6b556b01be31b8b14ffbc949802cfa05283beb90f40b7512c4b406d4b7ab30413a475bcf2362bbdb35a8ab7ef3f8a7e4ed7bd47615d7ff616f65f7a2ff26fa3530af68f85b6a760c347377a429d2e9fa021082e611e3b47912a3c1348d5065817f34b5d0696fe57b652830c671bc600891965d75601e1a81002503e152f745b3d058fb6f0a635050584e1c1c80a65e4e4893d47f36dbe35e9f97438277080f89adf5c0ddc3f111bc72627a235e49f3a1c78f52a0777cf2e2153b0ca695e5653122a915b454fe9da577e73f3a3c9778b4ecc3b61f61e27a7eaeb19b794c71ebc4dc13a8ff4795f6333286d0f70dc8c5c54a71421f66d047f7030d3977c6bc803eeccc9c94e60fe595164ac18ff932326b4392a7fc371608b723ef15df31248b871cbefa9a3656b79a97e75ac827220627098b6d39156bfd0e8aeb2f70d8b9f9f2e1d39815ef7396a5b217c32b26c5231b28f3d8ca13d638cdaf1b0d9593a21ae8806af6800aecc27a9d3127322011646c76cfea05b85983ebfc4efa1ebfe25eaddc5534c673e260dd6bfb55640219598f7e4e8bdc9b68585d89a94cc63f91a3e0d183a4c6bcc63b00ca55d16dd48fd43feb35f2fde9a947c2ebfc5d16453766d8531d1fbff8366f316b872a3101c012ce76674e07e51ed488b5cdc9cce3863931a63307c8ce7eb4fac414b3a2c74b6c2390b3bef9232cad5037019af9d700edcf3de0fc5dd64c66b65b8ad62db6d8c1af3301545798b5e34db2caf9befbd2383025eb7fcba5bf314b90ae2f02792a18df4c085d6934d0462e017002ab91520be6e0b176ca8249c9925b68dcc08df9731841863d9ecc492f6075e4e0aca3234a1ec30a4224068837b7fc624858b8fa81c63801e2b20b0f1ebe770717630000646ca9d2d297e3c56004fa5d9fe6bd9352bc7ad72e4554192c5bfe20d1d80ceb45bcb22c59d1dfa95351791ef217e6b6bd21c19a91a263f6f070d1b4eba75fc227408232b895b14b247c165b210b740a9ec6203a9703be91fa2b0cfab1cccacbdea8b5d6fd70eba8d100fc0393a3bd8fa4868d2539caf28f222528b238b26e160490a248466a10a073f24f952edfd3ae244812052f0346d63d149d2ad11fcb70c8ddf530477547e3379741e972934a63b2857a901df25104d69c149ff8f26f9978655dc2247492f0411e2e67c100e8739ec0a8b037062f0ffd63cf9d30a6504cca9cbf7ec7dffc93014308a1d638924ad09cd07668",
+        key: "9d29e43f7134c989f5b40ffb61bb4778c8fa95d9f8b27e4b0c2db33bc55a602c180550d2366b3c4f6a181bb83b046c214201f098a41baeb06ef5e95327aa7373486cd54af721ad47593bc95e9c2055f02f41492fc762e7dcf5e5e9010036a188c1b7023b59f1f6563a7ff47e1399a70ae05ab4b9203a7aa2e1c0030fb143b6d96ab3907a07c50ec9d69ba7b61a2c4038b49ffed3711b827dc683cf1b27ff39742b91911d7d45beba5ecea944d96c6397aa8f489bf875c64b57a8dbc543b23c3d23d5d34b6dd37b95ba7a537f83f45fb395b0275049b61a85fd5956f593d8fc3c32497c556479d52a277b5f93d573e07ef63675576ab9c8e84389dd82929bb259613b615b72e924a0c7715479a809b1541f8d0380b4dfe535d4b8c2bdc9440fa3858c3434c01a0b4b16a3a83c589d9db9b1bcde2438cd5f0a244e7879640ec51f18ce1d7219fb50ff843f1073d7f33fc421314cc3284e7d21a8fa161fa8181b1283645e78ab27dcee0965458ac8de4cac5558359c5d78c330146bfa",
+        mac: "11008133980ee7fad23844028a6a367baa9733e4f3e9b47c03b94b35fa99648c0bc4706b232728d8b76149c985a6a547d36f9349afb0b2043a5227b64e45337a",
+    },
+]
